+++ iqm-client/src/iqm/iqm_client/iqm_client.py
@@ -20,6 +20,7 @@ from functools import lru_cache
 from http import HTTPStatus
 from importlib.metadata import version
 import json
+import logging
 import os
 import platform
 import time
@@ -75,6 +76,9 @@ T_BaseModel = TypeVar("T_BaseModel", bound=BaseModel)
 REQUESTS_TIMEOUT = float(os.environ.get("IQM_CLIENT_REQUESTS_TIMEOUT", 120.0))
 DEFAULT_TIMEOUT_SECONDS = 900
 SECONDS_BETWEEN_CALLS = float(os.environ.get("IQM_CLIENT_SECONDS_BETWEEN_CALLS", 1.0))
+
+
+logger = logging.getLogger(__name__)
 
 
 class IQMClient:
@@ -907,10 +911,17 @@ class IQMClient:
             # model = model_class.model_validate_json(response.text)
         except json.decoder.JSONDecodeError as e:
             raise EndpointRequestError(f"Invalid response: {response.text}, {e!r}") from e
+
         return model
 
-    def get_structured_metrics(self, calibration_set_id: UUID | None = None) -> ObservationFinder:
+    def get_calibration_quality_metrics(self, calibration_set_id: UUID | None = None) -> ObservationFinder:
         """Retrieve the given calibration set and related quality metrics from the server.
+
+        .. warning::
+
+           This method is an experimental interface to the quality metrics and calibration data.
+           The API may change considerably in the next versions *with no backwards compatibility*,
+           including the API of the ObservationFinder class.
 
         Args:
             calibration_set_id: ID of the calibration set to retrieve.
@@ -925,8 +936,16 @@ class IQMClient:
             HTTPException: HTTP exceptions
 
         """
+        logger.warning(
+            "IQMClient.get_calibration_quality_metrics is an experimental method, and the API will likely change "
+            "in the future with no backwards compatibility."
+        )
+        return self._get_calibration_quality_metrics(calibration_set_id)
+
+    def _get_calibration_quality_metrics(self, calibration_set_id: UUID | None = None) -> ObservationFinder:
+        """See :meth:`get_calibration_quality_metrics`."""
         if isinstance(self._station_control, IqmServerClient):
-            raise ValueError("The get_structured_metrics method is not supported by IqmServerClient.")
+            raise ValueError("The _get_calibration_quality_metrics method is not supported by IqmServerClient.")
 
         if not calibration_set_id:
             # find out the default calset id
@@ -936,5 +955,4 @@ class IQMClient:
         calset_obs = self._station_control.get_observation_set_observations(calibration_set_id)
         quality_metrics = self._station_control.get_calibration_set_quality_metrics(calibration_set_id)
         qm_obs = quality_metrics.observations
-
-        return ObservationFinder(list(calset_obs) + qm_obs)
+        return ObservationFinder(calset_obs + qm_obs)